testing/20151210.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20151210</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<h1 id="w3c/testing.20151210.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2015-12-10 / <a href="#end">end</a></h1>
<p>Options: 
<p><a href=20151209>Previous day</a>, <a href=20151211>Next day</a></p>
<ol id=lines>
<li id=l-1 class=hide><a href=#l-1>#</a> Session Start: Thu Dec 10 00:00:00 2015</li>
<li id=l-2 class=hide><a href=#l-2>#</a> Session Ident: #testing</li>
<li id=l-3 class=hide><a href=#l-3>#</a> [00:16] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-4 class=hide _a="heycam"><a href=#l-4>#</a> [00:33] * heycam is now known as heycam|away</li>
<li id=l-5 class=hide><a href=#l-5>#</a> [00:35] * Joins: majidvp (~sid96638@public.cloak)</li>
<li id=l-6 class=hide _a="heycam|away"><a href=#l-6>#</a> [00:42] * heycam|away is now known as heycam|away|away</li>
<li id=l-7 class=hide><a href=#l-7>#</a> [00:46] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-8 class=hide><a href=#l-8>#</a> [01:01] * Joins: rniwa_ (~textual@public.cloak)</li>
<li id=l-9 class=hide><a href=#l-9>#</a> [01:06] * Quits: rniwa (~textual@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-10 class=hide><a href=#l-10>#</a> [01:59] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-11 class=hide><a href=#l-11>#</a> [02:10] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-12 class=hide><a href=#l-12>#</a> [02:16] * Quits: rniwa_ (~textual@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-13 class=hide _a="terri"><a href=#l-13>#</a> [02:41] * terri is now known as terri_offline</li>
<li id=l-14 class=hide><a href=#l-14>#</a> [02:45] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-15 class=hide><a href=#l-15>#</a> [02:45] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-16 class=hide><a href=#l-16>#</a> [02:45] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-17 class=hide><a href=#l-17>#</a> [02:46] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-18 class=hide><a href=#l-18>#</a> [02:49] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-19 class=hide><a href=#l-19>#</a> [02:52] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-20 class=hide><a href=#l-20>#</a> [02:53] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-21 class=hide><a href=#l-21>#</a> [02:56] * Joins: rniwa_ (~textual@public.cloak)</li>
<li id=l-22 class=hide><a href=#l-22>#</a> [02:57] * Quits: rniwa (~textual@public.cloak) (Client closed connection)</li>
<li id=l-23 class=hide><a href=#l-23>#</a> [03:00] * Quits: bhill2 (~bhill2@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-24 class=hide><a href=#l-24>#</a> [04:34] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-25 class=hide><a href=#l-25>#</a> [05:10] * Quits: frewsxcv (~sid86219@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-26 class=hide><a href=#l-26>#</a> [05:11] * Joins: frewsxcv (~sid86219@public.cloak)</li>
<li id=l-27 class=hide><a href=#l-27>#</a> [07:08] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-28 class=hide><a href=#l-28>#</a> [07:11] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-29 class=hide><a href=#l-29>#</a> [07:46] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-30 class=hide><a href=#l-30>#</a> [07:46] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-31 class=hide><a href=#l-31>#</a> [07:51] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-32 class=hide><a href=#l-32>#</a> [07:51] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-33 class=hide><a href=#l-33>#</a> [07:57] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-34 class=hide><a href=#l-34>#</a> [08:02] * Quits: bhill2 (~bhill2@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-35 class=hide><a href=#l-35>#</a> [08:32] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-36 class=hide><a href=#l-36>#</a> [08:49] * Quits: rniwa_ (~textual@public.cloak) (&quot;My Mac has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-37 class=hide><a href=#l-37>#</a> [09:41] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-38><a href=#l-38>#</a> [09:41] -gitbot:#testing- [csswg-test] hakatashi opened pull request #979: Revert 55a4412 for white-space tests (master...revert-white-spaces-test) <a href="https://github.com/w3c/csswg-test/pull/979">https://github.com/w3c/csswg-test/pull/979</a></li>
<li id=l-39 class=hide><a href=#l-39>#</a> [09:41] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-40 class=hide><a href=#l-40>#</a> [09:42] * Joins: rego (~rego@public.cloak)</li>
<li id=l-41 class=hide><a href=#l-41>#</a> [09:58] * Quits: rego (~rego@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-42 _a="mkwst"><a href=#l-42>#</a> [10:28] &lt;mkwst&gt; Hello #testing. I'd love some help finding someone to review <a href="https://github.com/w3c/web-platform-tests/pull/2385">https://github.com/w3c/web-platform-tests/pull/2385</a>.</li>
<li id=l-43 _a="zcorpan" _r="mkwst"><a href=#l-43>#</a> [11:21] &lt;zcorpan&gt; mkwst: i suppose this should affect websockets also?</li>
<li id=l-44 _a="mkwst" _r="zcorpan"><a href=#l-44>#</a> [11:21] &lt;mkwst&gt; zcorpan: The cookie change? Yes.</li>
<li id=l-45 _a="mkwst"><a href=#l-45>#</a> [11:23] &lt;mkwst&gt; Thanks for taking a look. :)</li>
<li id=l-46 _a="zcorpan"><a href=#l-46>#</a> [11:31] &lt;zcorpan&gt; thank you for fixing cookies :-)</li>
<li id=l-47 _a="mkwst" _r="zcorpan"><a href=#l-47>#</a> [12:35] &lt;mkwst&gt; zcorpan: Do you happen to know if I need to do anything special to register a new websocket handler?</li>
<li id=l-48 _a="zcorpan" _r="mkwst"><a href=#l-48>#</a> [12:38] &lt;zcorpan&gt; mkwst: it uses py_websocket. i think it should be enough to create a new file in <a href="https://github.com/w3c/web-platform-tests/tree/master/websockets/handlers">https://github.com/w3c/web-platform-tests/tree/master/websockets/handlers</a></li>
<li id=l-49 _a="mkwst" _r="zcorpan"><a href=#l-49>#</a> [12:39] &lt;mkwst&gt; zcorpan: Yeah, I assumed so as well, but it's giving me 404 errors. I'm sure I'm doing something wrong. :)</li>
<li id=l-50 _a="zcorpan" _r="mkwst"><a href=#l-50>#</a> [12:39] &lt;zcorpan&gt; mkwst: check what <a href="https://github.com/w3c/web-platform-tests/blob/master/websockets/cookies/001.html">https://github.com/w3c/web-platform-tests/blob/master/websockets/cookies/001.html</a> does</li>
<li id=l-51 _a="zcorpan" _r="mkwst"><a href=#l-51>#</a> [12:41] &lt;zcorpan&gt; mkwst: in particular the path should not include /websockets/handlers/ in the ws: or wss: URL</li>
<li id=l-52 _a="mkwst"><a href=#l-52>#</a> [12:41] &lt;mkwst&gt; Ah. I need to kill the server and start it up again.</li>
<li id=l-53 _a="zcorpan" _r="yeah"><a href=#l-53>#</a> [12:42] &lt;zcorpan&gt; yeah, that too</li>
<li id=l-54 _a="zcorpan"><a href=#l-54>#</a> [12:43] &lt;zcorpan&gt; the documentation doesn't cover websocket tests it seems :-(</li>
<li id=l-55 _a="mkwst"><a href=#l-55>#</a> [12:47] &lt;mkwst&gt; I'm noticing. But hey, cargo-culting mostly works.</li>
<li id=l-56 _a="zcorpan"><a href=#l-56>#</a> [12:48] &lt;zcorpan&gt; filed a bug. maybe i can fix it someday if someone reminds me to :-)</li>
<li id=l-57 _a="mkwst"><a href=#l-57>#</a> [12:49] &lt;mkwst&gt; can you add `assert_no_regexp_match` at the same time? :)</li>
<li id=l-58 class=hide><a href=#l-58>#</a> [12:52] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-59><a href=#l-59>#</a> [12:52] -gitbot:#testing- [web-platform-tests] mikewest pushed 1 new commit to mkwst/draft-west-leave-secure-cookies-alone: <a href="https://github.com/w3c/web-platform-tests/commit/6d58316b3377b36dfc20c77942e5abcc11582cf3">https://github.com/w3c/web-platform-tests/commit/6d58316b3377b36dfc20c77942e5abcc11582cf3</a></li>
<li id=l-60><a href=#l-60>#</a> [12:52] -gitbot:#testing- web-platform-tests/mkwst/draft-west-leave-secure-cookies-alone 6d58316 Mike West: fixup websockets + zcorpan</li>
<li id=l-61 class=hide><a href=#l-61>#</a> [12:52] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-62 _a="zcorpan"><a href=#l-62>#</a> [12:52] &lt;zcorpan&gt; that would be assert_not_regexp_match, or possibly assert_regexp_not_match</li>
<li id=l-63 _a="mkwst"><a href=#l-63>#</a> [12:53] &lt;mkwst&gt; Name it whatever you like. ;)</li>
<li id=l-64 _a="mkwst"><a href=#l-64>#</a> [12:53] &lt;mkwst&gt; I added some websocket tests. Mind taking another look while I grab lunch? :)</li>
<li id=l-65 class=hide><a href=#l-65>#</a> [13:04] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-66><a href=#l-66>#</a> [13:04] -gitbot:#testing- [testharness.js] zcorpan created assert_not_regexp_match (+1 new commit): <a href="https://github.com/w3c/testharness.js/commit/25f65333cd40">https://github.com/w3c/testharness.js/commit/25f65333cd40</a></li>
<li id=l-67><a href=#l-67>#</a> [13:04] -gitbot:#testing- testharness.js/assert_not_regexp_match 25f6533 Simon Pieters: Add assert_not_regexp_match...</li>
<li id=l-68 class=hide><a href=#l-68>#</a> [13:04] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-69 class=hide><a href=#l-69>#</a> [13:08] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-70><a href=#l-70>#</a> [13:08] -gitbot:#testing- [testharness.js] zcorpan opened pull request #167: Add assert_not_regexp_match (master...assert_not_regexp_match) <a href="https://github.com/w3c/testharness.js/pull/167">https://github.com/w3c/testharness.js/pull/167</a></li>
<li id=l-71 class=hide><a href=#l-71>#</a> [13:08] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-72 _a="zcorpan" _r="mkwst"><a href=#l-72>#</a> [13:12] &lt;zcorpan&gt; mkwst: have to leave now. will try to remember to check tomorrow. otherwise ping me</li>
<li id=l-73 _a="mkwst"><a href=#l-73>#</a> [13:12] &lt;mkwst&gt; No worries. Thanks for the review thus far!</li>
<li id=l-74 _a="jgraham"><a href=#l-74>#</a> [13:15] &lt;jgraham&gt; What are you using the not_regexp_match for?</li>
<li id=l-75 _a="jgraham"><a href=#l-75>#</a> [13:15] &lt;jgraham&gt; In general it seems like a very weak test, but I guess in some cases it could be OK</li>
<li id=l-76 _a="zcorpan" _r="jgraham"><a href=#l-76>#</a> [13:19] &lt;zcorpan&gt; jgraham: cookies</li>
<li id=l-77 _a="zcorpan"><a href=#l-77>#</a> [13:20] &lt;zcorpan&gt; i suppose it really wants to check that a substring is not present, but regexp handles that and more</li>
<li id=l-78 _a="zcorpan"><a href=#l-78>#</a> [13:20] &lt;zcorpan&gt; actually leaving now</li>
<li id=l-79 class=hide><a href=#l-79>#</a> [13:20] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-80 _a="jgraham"><a href=#l-80>#</a> [13:23] &lt;jgraham&gt; OK. I am just a little nervous about encouraging people to write tests where there are an infinite number of pass conditions and a small number of fail conditions</li>
<li id=l-81 class=hide><a href=#l-81>#</a> [14:01] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-82 class=hide><a href=#l-82>#</a> [14:01] * Joins: dom (dom@public.cloak)</li>
<li id=l-83 class=hide><a href=#l-83>#</a> [14:21] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-84 class=hide><a href=#l-84>#</a> [14:28] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-85 _a="mkwst" _r="jgraham"><a href=#l-85>#</a> [14:36] &lt;mkwst&gt; jgraham: Well, in this particular case, I'm verifying that a particular cookie doesn't exist. Since that boils down to a string match (hooray, `document.cookies`!), this seemed like a reasonable approach.</li>
<li id=l-86 _a="mkwst" _r="jgraham"><a href=#l-86>#</a> [14:37] &lt;mkwst&gt; jgraham: If there's a different way you'd like me to write <a href="https://github.com/w3c/web-platform-tests/pull/2385/files#diff-41a3813a440607785b744b25f650d221R26">https://github.com/w3c/web-platform-tests/pull/2385/files#diff-41a3813a440607785b744b25f650d221R26</a>, I'm happy to change it.</li>
<li id=l-87 class=hide><a href=#l-87>#</a> [14:38] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-88 class=hide _a="heycam|away|away"><a href=#l-88>#</a> [14:59] * heycam|away|away is now known as heycam|away</li>
<li id=l-89 class=hide><a href=#l-89>#</a> [14:59] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-90 class=hide><a href=#l-90>#</a> [15:00] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-91 class=hide _a="heycam|away"><a href=#l-91>#</a> [15:00] * heycam|away is now known as heycam</li>
<li id=l-92 class=hide><a href=#l-92>#</a> [15:03] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-93><a href=#l-93>#</a> [15:03] -gitbot:#testing- [csswg-test] Ms2ger closed pull request #979: Revert &quot;Remove trailing whitespace&quot; for white-space tests (master...revert-white-spaces-test) <a href="https://github.com/w3c/csswg-test/pull/979">https://github.com/w3c/csswg-test/pull/979</a></li>
<li id=l-94 class=hide><a href=#l-94>#</a> [15:03] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-95 class=hide><a href=#l-95>#</a> [15:13] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-96 class=hide><a href=#l-96>#</a> [15:21] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-97 class=hide><a href=#l-97>#</a> [15:39] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-98 _a="jgraham" _r="mkwst"><a href=#l-98>#</a> [15:43] &lt;jgraham&gt; mkwst: Worst API Ever</li>
<li id=l-99 _a="jgraham"><a href=#l-99>#</a> [15:43] &lt;jgraham&gt; (well except all the other worst ones)</li>
<li id=l-100 _a="mkwst"><a href=#l-100>#</a> [15:43] &lt;mkwst&gt; There are lots of bad APIs. Cookies is up there.</li>
<li id=l-101 _a="jgraham" _r="Unrelatedly"><a href=#l-101>#</a> [15:45] &lt;jgraham&gt; Unrelatedly, I was going to implement the webdriver cookies support in servo recently until I realised that it wants a structured view of cookies and getting anything other than a document.cookies-like string representation from the existing servo APIs would be more work than I had time available right then</li>
<li id=l-102 _a="jgraham" _r="But"><a href=#l-102>#</a> [15:47] &lt;jgraham&gt; But, yeah, I guess I can tolerate using a regexp in that case. But I think I'm -0.5 on adding regexp_not_match to testharness.js because I think it's not something you should do very often, and when you do have to do it it's easy enough</li>
<li id=l-103 _a="mkwst" _r="jgraham"><a href=#l-103>#</a> [16:02] &lt;mkwst&gt; jgraham: As long as you know it doesn't exist, sure. I assumed it did, and wasted a minute or three tracking down my mistake.</li>
<li id=l-104 _a="mkwst"><a href=#l-104>#</a> [16:03] &lt;mkwst&gt; *shrug*</li>
<li id=l-105 class=hide><a href=#l-105>#</a> [16:10] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-106 class=hide><a href=#l-106>#</a> [16:34] * Joins: abreton (~abreton@public.cloak)</li>
<li id=l-107 class=hide><a href=#l-107>#</a> [16:39] * Joins: hgl_ (~hgl@public.cloak)</li>
<li id=l-108 class=hide><a href=#l-108>#</a> [16:41] * Quits: hgl (~hgl@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-109 class=hide><a href=#l-109>#</a> [16:46] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-110 class=hide><a href=#l-110>#</a> [16:54] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-111 class=hide><a href=#l-111>#</a> [17:17] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-112 class=hide><a href=#l-112>#</a> [17:27] * Quits: abreton (~abreton@public.cloak) (&quot;Page closed&quot;)</li>
<li id=l-113 class=hide><a href=#l-113>#</a> [17:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-114><a href=#l-114>#</a> [17:29] -gitbot:#testing- [testharness.js] zcorpan closed pull request #167: Add assert_not_regexp_match (master...assert_not_regexp_match) <a href="https://github.com/w3c/testharness.js/pull/167">https://github.com/w3c/testharness.js/pull/167</a></li>
<li id=l-115 class=hide><a href=#l-115>#</a> [17:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-116 class=hide><a href=#l-116>#</a> [17:29] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-117 class=hide><a href=#l-117>#</a> [17:30] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-118 class=hide _a="heycam"><a href=#l-118>#</a> [17:30] * heycam is now known as heycam|away</li>
<li id=l-119 class=hide><a href=#l-119>#</a> [17:30] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-120 class=hide><a href=#l-120>#</a> [17:30] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-121 class=hide><a href=#l-121>#</a> [17:36] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-122 class=hide><a href=#l-122>#</a> [17:38] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-123 class=hide><a href=#l-123>#</a> [18:02] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-124 class=hide><a href=#l-124>#</a> [18:03] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-125 class=hide><a href=#l-125>#</a> [18:03] * Quits: fwtnb (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-126 class=hide><a href=#l-126>#</a> [18:03] * Joins: fwtnb_ (~fwtnb@public.cloak)</li>
<li id=l-127 class=hide><a href=#l-127>#</a> [18:15] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-128 class=hide><a href=#l-128>#</a> [18:17] * Joins: ted (ted@public.cloak)</li>
<li id=l-129 class=hide><a href=#l-129>#</a> [18:30] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-130 class=hide><a href=#l-130>#</a> [18:40] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-131 class=hide _a="terri_offline"><a href=#l-131>#</a> [18:55] * terri_offline is now known as terri</li>
<li id=l-132 class=hide><a href=#l-132>#</a> [18:56] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-133 class=hide><a href=#l-133>#</a> [18:56] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-134 class=hide><a href=#l-134>#</a> [19:01] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-135 class=hide><a href=#l-135>#</a> [19:03] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-136 class=hide><a href=#l-136>#</a> [19:04] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-137 class=hide><a href=#l-137>#</a> [19:06] * Joins: bhill2_ (~bhill2@public.cloak)</li>
<li id=l-138 class=hide><a href=#l-138>#</a> [19:06] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-139 class=hide><a href=#l-139>#</a> [19:20] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-140 class=hide><a href=#l-140>#</a> [19:24] * Quits: fwtnb_ (~fwtnb@public.cloak) (Client closed connection)</li>
<li id=l-141 class=hide _a="heycam|away"><a href=#l-141>#</a> [19:38] * heycam|away is now known as heycam</li>
<li id=l-142 class=hide><a href=#l-142>#</a> [19:39] * Quits: Ms2ger (~Ms2ger@public.cloak) (Client closed connection)</li>
<li id=l-143 class=hide><a href=#l-143>#</a> [19:39] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-144 class=hide><a href=#l-144>#</a> [20:23] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-145><a href=#l-145>#</a> [20:23] -gitbot:#testing- [csswg-test] Ms2ger closed pull request #969: Move the DOMPoint test a little closer to the specification. (master...point) <a href="https://github.com/w3c/csswg-test/pull/969">https://github.com/w3c/csswg-test/pull/969</a></li>
<li id=l-146 class=hide><a href=#l-146>#</a> [20:23] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-147 class=hide><a href=#l-147>#</a> [20:28] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-148 class=hide><a href=#l-148>#</a> [20:36] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-149 class=hide><a href=#l-149>#</a> [20:45] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-150 class=hide><a href=#l-150>#</a> [20:46] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-151 class=hide><a href=#l-151>#</a> [20:46] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-152 class=hide><a href=#l-152>#</a> [21:04] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-153 class=hide><a href=#l-153>#</a> [21:05] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-154 class=hide><a href=#l-154>#</a> [21:07] * Joins: rniwa (~textual@public.cloak)</li>
<li id=l-155 class=hide><a href=#l-155>#</a> [21:12] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-156 class=hide><a href=#l-156>#</a> [21:31] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-157 class=hide><a href=#l-157>#</a> [21:31] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-158 class=hide><a href=#l-158>#</a> [21:31] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-159 class=hide><a href=#l-159>#</a> [21:38] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-160 class=hide><a href=#l-160>#</a> [21:55] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-161 class=hide><a href=#l-161>#</a> [21:58] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-162 class=hide _a="heycam"><a href=#l-162>#</a> [21:58] * heycam is now known as heycam|away</li>
<li id=l-163 class=hide><a href=#l-163>#</a> [22:04] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-164 class=hide><a href=#l-164>#</a> [22:09] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-165 class=hide><a href=#l-165>#</a> [22:13] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
<li id=l-166 _a="ato" _r="mkwst"><a href=#l-166>#</a> [22:22] &lt;ato&gt; mkwst: Cookies is definitely up there.  I’ve been spending the last month deciphering the specification(s).</li>
<li id=l-167 class=hide><a href=#l-167>#</a> [22:23] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-168 class=hide><a href=#l-168>#</a> [22:25] * Joins: fwtnb (~fwtnb@public.cloak)</li>
<li id=l-169 class=hide><a href=#l-169>#</a> [22:32] * Quits: fwtnb (~fwtnb@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-170 class=hide><a href=#l-170>#</a> [22:46] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-171 class=hide _a="heycam|away"><a href=#l-171>#</a> [23:13] * heycam|away is now known as heycam</li>
<li id=l-172 class=hide _a="heycam"><a href=#l-172>#</a> [23:20] * heycam is now known as heycam|away</li>
<li id=l-173 class=hide _a="heycam|away"><a href=#l-173>#</a> [23:27] * heycam|away is now known as heycam</li>
<li id=l-174 class=hide><a href=#l-174>#</a> Session Close: Fri Dec 11 00:00:00 2015</li>
</ol>
<p id=end><a href=20151209>Previous day</a>, <a href=20151211>Next day</a></p>
<p>Think <a href="/irc-logs/">these logs</a> are useful? Then <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&amp;business=PMXKLAEH53JSS&amp;lc=GB&amp;item_name=Krijn%20Hoetmer&amp;item_number=IRC%20Logs&amp;currency_code=EUR&amp;bn=PP%2dDonationsBF%3abtn_donate_SM%2egif%3aNonHosted"><strong>please donate</strong></a> to show your gratitude (and keep them up, of course). <a href="https://twitter.com/krijnhoetmer/status/414824591337734144">Thanks! — Krijn</a></p>
<script src=/irc-logs/s2.js></script>
